workspace/20250812_003905_ocr_poc/code/requirements.txt
easyocr==1.4.1
Pillow==8.4.0
numpy==1.21.2
pandas==1.3.3
matplotlib==3.4.3